Extracted layout aspects of RulesDrawer to RightDrawer.
