+++ frontend/src/components/RightDrawer/index.spec.tsx
@@ -1,0 +1,8 @@
+import '@testing-library/jest-dom/extend-expect';
+import { RightDrawer } from '.';
+
+describe('RightDrawer', () => {
+  it('is defined', () => {
+    expect(RightDrawer).toBeDefined();
+  });
+});
