--- frontend/src/pages/registry/WmsSecuritySettings/components/RulesDrawer/index.spec.tsx
@@ -1,8 +1,0 @@
-import '@testing-library/jest-dom/extend-expect';
-import { RulesDrawer } from '.';
-
-describe('RulesTable', () => {
-  it('is defined', () => {
-    expect(RulesDrawer).toBeDefined();
-  });
-});
