+++ frontend/src/pages/registry/WmsSecuritySettings/components/RulesDrawer/index.tsx
@@ -1,10 +1,7 @@
-import { LeftOutlined, RightOutlined } from '@ant-design/icons';
-import { Button, Drawer } from 'antd';
+import { RightDrawer } from '@/components/RightDrawer';
 import type OlMap from 'ol/Map';
 import type { ReactElement } from 'react';
-import { useEffect, useRef, useState } from 'react';
 import { useLocation } from 'react-router';
-import './index.css';
 import { RuleForm } from './RuleForm/RuleForm';
 import { RulesTable } from './RulesTable/RulesTable';
 
@@ -23,26 +20,7 @@ export const RulesDrawer = ({
   setIsRuleEditingActive,
   map,
 }: RulesDrawerProps): ReactElement => {
-  const buttonRef = useRef<HTMLButtonElement>(null);
-  const [isVisible, setIsVisible] = useState<boolean>(true);
   const path = useLocation().pathname;
-
-  // adjust padding of map div
-  useEffect(() => {
-    if (map) {
-      const mapDiv: any = document.querySelector(`#${map.getTarget()}`);
-      if (!isVisible) {
-        mapDiv.style.paddingRight = '0px';
-      } else {
-        mapDiv.style.paddingRight = '500px';
-      }
-    }
-  }, [map, isVisible]);
-
-  const toggleVisible = () => {
-    setIsVisible(!isVisible);
-    buttonRef.current?.blur();
-  };
 
   let content;
   if (path.endsWith('edit')) {
@@ -66,25 +44,5 @@ export const RulesDrawer = ({
   } else {
     content = <RulesTable wmsId={wmsId} />;
   }
-  return (
-    <>
-      <Button
-        ref={buttonRef}
-        className={`rules-drawer-toggle-button ${isVisible ? 'expanded' : 'collapsed'}`}
-        onClick={toggleVisible}
-        icon={isVisible ? <RightOutlined /> : <LeftOutlined />}
-      />
-      <Drawer
-        placement="right"
-        getContainer={false}
-        width={500}
-        visible={isVisible}
-        closable={false}
-        mask={false}
-        style={{ zIndex: 1, height: '100%', marginTop: '48px' }}
-      >
-        {content}
-      </Drawer>
-    </>
-  );
+  return <RightDrawer map={map}>{content}</RightDrawer>;
 };
